requirements.txt
python-dotenv
opensearch-py
openai
rapidfuzz
Flask
gunicorn
flask-cors
